Fix: correctness issue in optimal solution (refinement).
After refining the tree, some of the arcs in the DD are removed. It
is possilbe that after removing some of the arcs, the DD tree ends up
as a small forest of non-connected trees. To fix this, we added
additional code to check the status of tree after removing arcs.

Revert to previous commit it this commit breaks the remote branch.