Add: reusing gurobi model for all subproblems.
When solving a subproblem, a new gurobi model is created and all
the gurobi variables are initialized. The variables and the constraints
remains same for all the scenarios and subproblems. Only the objective
function changes for each scenario (only the coefficients).

Reusing the model and the variables might provide performance by saving
the cost of memory allocation for gurobi variables. This might not get
significant performance if memory allocator is smart enough to recognize
the pattern and not returning the memory to the heap.

This feature is only tested on local instances. Revert to previous commit
if issue appears with correctness.